requirements.txt
django
gunicorn
django_on_heroku
whitenoise
beautifulsoup4
selenium